requirements.txt
Pillow>=9.0.0
pyinstaller>=5.0
